Split up/rename the resource and providers
This is preparation for any `dropbox_config` and/or `dropbox_service`
resources that may be added in the future.
